requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
pip==24.2
pluggy==1.5.0
pytest==8.3.3
requests==2.32.3
rpds-py==0.23.1
typing_extensions==4.12.2
urllib3==2.2.3
